9397: Factor out FiltersModel
